+++ server/src/logic/sync/presentation/routes.ts
@@ -53,6 +53,9 @@ export class SyncFileSystemRoutes {
             id: sync.id,
             title: sync.title,
             connector: sync.connector.name,
+            lastSyncGMT: sync.lastSyncGMT,
+            disabled: sync.disabled,
+            totalSyncedResources: sync.originalIds?.length || 0,
           }));
         res.status(200).send(kbSyncs);
       } catch (error) {
@@ -87,6 +90,8 @@ export class SyncFileSystemRoutes {
       try {
         await this.checkAuth(id, req.headers.token as string, syncRepository);
         const data = await new GetSync(syncRepository).execute(id);
+        // remove originalIds from response (too big)
+        delete data.originalIds;
         res.status(200).send(data);
       } catch (error) {
         this.handleError(res, error);